First Version
This is the base functionality to allow a prompt. The goal of the
library is to provide something extremely simple, so rather putting a
bunch of bells and whistles in I've focused it purely on providing a
prompt and processing the input.

I ended up debating the license choice quite a bit, ultimately
settling on the GPL v3. The primary motivation for choosing a less
restrictive license would be so more people could use the library, but
this provides such a simple utility if people really want to use it in
a proprietary context they can just reimplement it.